Use generics to pass `EncoderDecoder.Config` type (#75)

+++ library/encoding-core/src/commonTest/kotlin/io/matthewnelson/encoding/core/util/DecoderInputUnitTest.kt
@@ -41,7 +41,7 @@ class DecoderInputUnitTest {
 
         val actual = config.decodeOutMaxSizeOrFail(DecoderInput(encoded))
 
-        assertEquals(expectedOutSize.toInt(), actual)
+        assertEquals(expectedOutSize, actual)
     }
 
     @Test
